+++ GuaranteedRate.Sextant/Logging/File/FileLogAppender.cs
@@ -136,11 +136,7 @@ namespace GuaranteedRate.Sextant.Logging.File
                         {
                             return false;
                         }
-                        catch (Exception ex)
-                        {
-
-                            throw;
-                        }
+                       
                     }
                 }
             }