+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## v17.3.2.0 / 2017 Oct 06
 
 * **Fix**  -  Numerous fixes to logging.  
-* **Add** Expose Shutdown() on logger.  Handle recursive errors.
-* **Add** Prmitive file logger (mainly useful in debugging other logging failures).
+* **Add**  -  Expose Shutdown() on logger.  Handle recursive errors.
+* **Add**  -  Prmitive file logger (mainly useful in debugging other logging failures).
 
 ```csharp
 [GuaranteedRate.Sextant "17.3.2.0"]